Fix V2 segment format issues
1. Fix doc length truncation: Change TpDocMapEntry.doc_length from
   uint16 to uint32 to preserve precision when decoding/re-encoding
   fieldnorms during segment merge. Documents with >65535 terms were
   previously losing precision.

2. Add warnings for missing CTID in docmap: When tp_docmap_lookup
   returns UINT32_MAX (CTID not found), log a WARNING before falling
   back to doc_id 0. This should never happen in practice but if it
   does, the warning makes debugging easier.

3. Eliminate double hash lookups in segment writing: Restructure
   tp_write_segment_v2 and write_merged_segment to build block_postings
   once (including fieldnorm lookup) and reuse for both skip entry
   computation and posting block writing. This roughly halves the
   docmap hash lookups during segment creation.

4. Fix block_max_norm computation in merge: Previously hardcoded to 0,
   now properly computed from fieldnorms in the block.

+++ src/segment/docmap.c
@@ -70,7 +70,7 @@ tp_docmap_create(void)
 }
 
 uint32
-tp_docmap_add(TpDocMapBuilder *builder, ItemPointer ctid, uint16 doc_length)
+tp_docmap_add(TpDocMapBuilder *builder, ItemPointer ctid, uint32 doc_length)
 {
 	TpDocMapEntry *entry;
 	bool		   found;
